# Avoid unnecessary substring invocations while parsing sls versions