--- a/changelog/@unreleased/pr-488.v2.yml
+++ b/changelog/@unreleased/pr-488.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Avoid unnecessary substring invocations while parsing sls versions
+  links:
+  - https://github.com/palantir/sls-version-java/pull/488