--- a/versions.lock
+++ b/versions.lock
@@ -7,11 +7,7 @@ com.palantir.safe-logging:logger-slf4j:1.18.0 (1 constraints: 350e9350)
 com.palantir.safe-logging:logger-spi:1.18.0 (2 constraints: 791e0cb1)
 com.palantir.safe-logging:preconditions:1.18.0 (1 constraints: 3c05413b)
 com.palantir.safe-logging:safe-logging:1.18.0 (4 constraints: 52342904)
-net.sf.jopt-simple:jopt-simple:4.6 (1 constraints: 610a91b7)
-org.apache.commons:commons-math3:3.2 (1 constraints: 5c0a8ab7)
 org.immutables:value:2.8.8 (1 constraints: 14051536)
-org.openjdk.jmh:jmh-core:1.32 (1 constraints: 1411a7be)
-org.openjdk.jmh:jmh-generator-annprocess:1.32 (1 constraints: da04f730)
 org.slf4j:slf4j-api:1.7.31 (1 constraints: 471091b6)
 
 [Test dependencies]