@svc-autorelease 2021-08-09
Released [1.2.0](https://github.com/palantir/sls-version-java/releases/tag/1.2.0)