@carterkozak 2021-08-09
> I'd forgotten about that little method... how did you come across this?

Noticed it while updating the safe-logging PR #483 ;-)